+++ example/lib/src/screen/components/listview/listview_item_seperated_builder_screen.dart
@@ -14,7 +14,7 @@ class ListViewItemSeparatedBuilderScreen extends StatelessWidget {
         onRefresh: () async => Future.delayed(const Duration(seconds: 2)),
         separatorType: ImpaktfullSeparatorType.canvas,
         items: List.generate(100, (index) => 'Child ${index + 1}'),
-        itemBuilder: (context, item) => Text(item),
+        itemBuilder: (context, item, index) => Text(item),
         noDataLabel: 'No Data found',
       ),
     );